+++ Tests/FeatherWidgets.TestUI.Arrangements/Packaging/StaticContent/News/ImportEditedNewsModule.cs
@@ -49,7 +49,10 @@ namespace FeatherWidgets.TestUI.Arrangements
         {
             ServerOperations.News().DeleteAllNews();
             ServerOperations.Pages().DeleteAllPages();
-            ServerOperations.ModuleBuilder().DeleteDirectory(this.tempFolderPath);
+
+            if (System.IO.Directory.Exists(this.tempFolderPath))
+                ServerOperations.ModuleBuilder().DeleteDirectory(this.tempFolderPath);
+
             ServerOperations.Packaging().DeleteAllPackagesFromDB();
 
             for (int i = 0; i < this.widgetTemplatesNames.Length; i++)
